Fix bug __repr__ting an XmlModel on Python3 (doesn't define unicode)
The `__repr__` implementation on class `XmlModel` uses the function
`unicode` which is only defined in Python 2. So in Python 3, this
raises an exception.

I changed the usage of `unicode` with `smart_str` from
`django.utils.encoding` as it promises to use `smart_bytes` if
running in Python 2 and `smart_text` if Python 3.

What happens is that Python 3 expects `__repr__` to return a
`str` not a bytestring. That's why changed the import logic and
leverage on Django logic behind the `smart_*` methods handling
Python 2 and 3.